+++ bagua/torch_api/algorithms/bytegrad.py
@@ -53,6 +53,7 @@ class ByteGradAlgorithm_Implementation(Algorithm):
             average=self.average,
             scattergather=True,
             compression="MinMaxUInt8",
+            group=bagua_module._bagua_process_group,
         )