+++ bagua/torch_api/algorithms/gradient_allreduce.py
@@ -26,17 +26,11 @@ class GradientAllReduceAlgorithm_Implementation(Algorithm):
         bucket: BaguaBucket,
     ):
         bucket.clear_ops()
-        if self.hierarchical:
-            bucket.append_centralized_synchronous_op(
-                hierarchical=self.hierarchical,
-                average=self.average,
-            )
-        else:
-            bucket.append_centralized_synchronous_op(
-                hierarchical=self.hierarchical,
-                average=self.average,
-            )
-
+        bucket.append_centralized_synchronous_op(
+            hierarchical=self.hierarchical,
+            average=self.average,
+            group=bagua_module._bagua_process_group,
+        )
 
 class GradientAllReduceAlgorithm:
     def __init__(self, hierarchical: bool = False, average: bool = True):