+++ bagua/torch_api/algorithms/q_adam.py
@@ -174,6 +174,7 @@ class QAdamAlgorithm_Implementation(Algorithm):
             bucket.append_centralized_synchronous_op(
                 hierarchical=False,
                 average=True,
+                group=bagua_module._bagua_process_group,
             )
         else:
 
@@ -188,6 +189,7 @@ class QAdamAlgorithm_Implementation(Algorithm):
                 average=True,
                 scattergather=True,
                 compression="MinMaxUInt8",
+                group=bagua_module._bagua_process_group,
             )
 
     def init_backward_hook(self, bagua_module: BaguaModule):