+++ bagua/torch_api/model_parallel/moe/experts.py
@@ -17,13 +17,13 @@ class Experts(torch.nn.Module):
     def __init__(self, expert, num_local_experts=1):
         super(Experts, self).__init__()
 
-        self.deepspeed_experts = torch.nn.ModuleList(
+        self.bagua_experts = torch.nn.ModuleList(
             [copy.deepcopy(expert) for i in range(num_local_experts)]
         )
         self.num_local_experts = num_local_experts
 
         # TODO: revisit allreduce for moe.gate...
-        for expert in self.deepspeed_experts:
+        for expert in self.bagua_experts:
             # TODO: Create param groups to handle expert + data case (e.g. param.group = moe_group)
             for name, param in expert.named_parameters():
                 param.allreduce = False
@@ -31,7 +31,7 @@ class Experts(torch.nn.Module):
     def forward(self, inputs):
         chunks = inputs.chunk(self.num_local_experts, dim=1)
         expert_outputs = []
-        for chunk, expert in zip(chunks, self.deepspeed_experts):
+        for chunk, expert in zip(chunks, self.bagua_experts):
             out = expert(chunk)
             if type(out) is tuple:
                 out = out[0]  # Ignore the bias term for now